Move demo code out of VisualEditorPanel
